♻️ refactor: renderChart 함수 분리.
